requirements.txt
flask
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2
markdown
openai
tqdm
nltk
numpy
scikit-learn
